refactor: dml query 들 각각 분리하고 이름 변경
